+++ src/js/core/core.js
@@ -1,6 +1,7 @@
 (function(core) {
 
     if (typeof define == "function" && define.amd) { // AMD
+
         define("uikit", function(){
 
             var uikit = window.UIkit || core(window, window.jQuery, window.document);
@@ -42,7 +43,7 @@
 
     var UI = {}, _UI = global.UIkit ? Object.create(global.UIkit) : undefined;
 
-    UI.version = '2.20.3';
+    UI.version = '2.21.0';
 
     UI.noConflict = function() {
         // restore UIkit version
@@ -141,6 +142,10 @@
     UI.support.mutationobserver = (global.MutationObserver || global.WebKitMutationObserver || null);
 
     UI.Utils = {};
+
+    UI.Utils.isFullscreen = function() {
+        return document.webkitFullscreenElement || document.mozFullScreenElement || document.msFullscreenElement || document.fullscreenElement || false;
+    };
 
     UI.Utils.str2json = function(str, notevil) {
         try {
@@ -369,12 +374,12 @@
         });
     };
 
-    $.UIkit      = UI;
-    $.fn.uk      = UI.fn;
+    $.UIkit          = UI;
+    $.fn.uk          = UI.fn;
 
     UI.langdirection = UI.$html.attr("dir") == "rtl" ? "right" : "left";
 
-    UI.components = {};
+    UI.components    = {};
 
     UI.component = function(name, def) {